+++ database/goploy.sql
@@ -392,7 +392,7 @@ CREATE TABLE IF NOT EXISTS `terminal_log` (
 INSERT IGNORE INTO `user`(`id`, `account`, `password`, `name`, `contact`, `state`, `super_manager`) VALUES (1, 'admin', '$2a$10$89ZJ2xeJj35GOw11Qiucr.phaEZP4.kBX6aKTs7oWFp1xcGBBgijm', '超管', '', 1, 1);
 INSERT IGNORE INTO `namespace`(`id`, `name`) VALUES (1, 'goploy');
 INSERT IGNORE INTO `namespace_user`(`id`, `namespace_id`, `user_id`, `role_id`) VALUES (1, 1, 1, 0);
-INSERT IGNORE INTO `system_config` (`id`, `key`, `value`) VALUES (1, 'version', '1.16.3');
+INSERT IGNORE INTO `system_config` (`id`, `key`, `value`) VALUES (1, 'version', '1.17.0');
 INSERT IGNORE INTO `role`(`id`, `name`, `description`) VALUES (1, 'manager', '');
 INSERT IGNORE INTO `role`(`id`, `name`, `description`) VALUES (2, 'member', '');
 INSERT IGNORE INTO `permission`(`id`, `pid`, `name`, `sort`, `description`) VALUES (1, 0, 'Log', 0, '');
@@ -457,7 +457,6 @@ INSERT IGNORE INTO `permission`(`id`, `pid`, `name`, `sort`, `description`) VALU
 INSERT IGNORE INTO `permission`(`id`, `pid`, `name`, `sort`, `description`) VALUES (60, 56, 'DeployDetail', 0, '');
 INSERT IGNORE INTO `permission`(`id`, `pid`, `name`, `sort`, `description`) VALUES (61, 56, 'DeployProject', 0, '');
 INSERT IGNORE INTO `permission`(`id`, `pid`, `name`, `sort`, `description`) VALUES (62, 56, 'DeployResetState', 0, '');
-INSERT IGNORE INTO `permission`(`id`, `pid`, `name`, `sort`, `description`) VALUES (63, 56, 'GreyDeploy', 0, '');
 INSERT IGNORE INTO `permission`(`id`, `pid`, `name`, `sort`, `description`) VALUES (64, 56, 'DeployRollback', 0, '');
 INSERT IGNORE INTO `permission`(`id`, `pid`, `name`, `sort`, `description`) VALUES (65, 56, 'DeployReview', 0, '');
 INSERT IGNORE INTO `permission`(`id`, `pid`, `name`, `sort`, `description`) VALUES (66, 56, 'DeployTask', 0, '');
@@ -525,7 +524,6 @@ INSERT IGNORE INTO `role_permission`(`role_id`, `permission_id`) VALUES (1, 59);
 INSERT IGNORE INTO `role_permission`(`role_id`, `permission_id`) VALUES (1, 60);
 INSERT IGNORE INTO `role_permission`(`role_id`, `permission_id`) VALUES (1, 61);
 INSERT IGNORE INTO `role_permission`(`role_id`, `permission_id`) VALUES (1, 62);
-INSERT IGNORE INTO `role_permission`(`role_id`, `permission_id`) VALUES (1, 63);
 INSERT IGNORE INTO `role_permission`(`role_id`, `permission_id`) VALUES (1, 64);
 INSERT IGNORE INTO `role_permission`(`role_id`, `permission_id`) VALUES (1, 65);
 INSERT IGNORE INTO `role_permission`(`role_id`, `permission_id`) VALUES (1, 66);